move class methods to their own file
